% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_prop_ab.R
\name{bayes_prop_ab}
\alias{bayes_prop_ab}
\title{A Bayes probability calculation from two proportions}
\usage{
bayes_prop_ab(prior_alpha, prior_beta, p1, p2, n1, n2, trials)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This functions allows you to quickly calculate the probability that one proportion is greater than another
}
\examples{
bayes_prop_ab()
}
\keyword{ab}
\keyword{bayes,}
\keyword{proportion}
\keyword{test,}
