% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_prop_statistics.R
\name{one_prop_statistics}
\alias{one_prop_statistics}
\title{Generate Bayesian statistics for one proportion, includes test against a hypothetical value}
\usage{
one_prop_statistics(
  prior_alpha,
  prior_beta,
  proportion,
  sample_size,
  prop_to_test,
  lower_ci,
  upper_ci
)
}
\description{
This function allows you to quickly generate summary bayesian statistics, visualize the beta distribution, and compare against a hypothetical value
}
\examples{
visualize_two_props()
}
\keyword{ab}
\keyword{bayes,}
\keyword{proportion}
\keyword{test,}
