% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_prop_statistics.R
\name{two_prop_statistics}
\alias{two_prop_statistics}
\title{Generate Bayesian statistics for two proportions, calcualte probability that one proportion is greater than other}
\usage{
two_prop_statistics(
  prior_alpha,
  prior_beta,
  proportion,
  sample_size,
  prop_to_test,
  lower_ci,
  upper_ci
)
}
\description{
This function allows you to quickly generate summary bayesian statistics, visualize two distributions
}
\examples{
visualize_two_props()
}
\keyword{ab}
\keyword{bayes,}
\keyword{proportion}
\keyword{test,}
