% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_two_props.R
\name{visualize_two_props}
\alias{visualize_two_props}
\title{A visualization of two proportions using 5million samples}
\usage{
visualize_two_props(prior_alpha, prior_beta, p1, p2, n1, n2, trials)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to quickly visualize the sampled beta distributions from two proportions
}
\examples{
visualize_two_props()
}
\keyword{ab}
\keyword{bayes,}
\keyword{proportion}
\keyword{test,}
